Fix setting timing flag in client creation
Fixed the timing flag logic so that `timing` given at the time of client
creation is used as fallback in case the `timing` is not given in
request parameters.

Added nock as a dependency to test the logic since it requires a real
`request` implementation to be called together with a client.