Update Node types and TypeScript
Updated Node types to latest version 12 to reflect the dropping of Node
8 and Node 10 from support.

This required TypeScript to be updated since TypeScript < 3.7 is not
support starting from @types/node v12.20.12 (ref.
https://github.com/DefinitelyTyped/DefinitelyTyped/issues/55430). The
chosen TypeScript version was the latest 3.9 version. The update to
TypeScript 4.x can be done as a followup.

Updating the two libraries required some type changes in the
implementation, notably the possibly breaking change of changing the
type of `ServiceClientRequestOptions.query` from `object | undefined` to
`ParsedUrlQueryInput | undefined`, which is more specific about the
possible query parameter types.